secproto_admin_backend_final/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
bcrypt==4.2.0
python-jose[cryptography]==3.3.0
pydantic==2.9.2
